Update Box3.js
the order in which  isIntersectionBox and containsPoint was checking for overlapping was to see if the object in question was NOT overlapping, then returning false if evaluating as true, and true if evaluating as false. That's really counter intuitive and confusing. I switched the less than or grater than signs and the false trues so that if the function evaluates as true, it returns true, if not it returns false. This should't change the way users interact with the function but it's much more intuitive source code.